{"types":"C","runs":4}

+++ Problem15.java
@@ -2,8 +2,8 @@ import acm.program.*;
 
 public class Problem15 extends ConsoleProgram {
 	public void run() {
-		int x = readInt("Enter the first number: ");
-		int y = readInt("Enter the first number: ");
+		double x = readInt("Enter the first number: ");
+		double y = readInt("Enter the first number: ");
 		println("the arithmetic mean of these numbers is: " + (double) ((x + y) / 2));
 	}
 }
